requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
PyPDF2==3.0.1
openai==1.12.0
#for connection to azure ai search:
azure-search-documents
